% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reading RBA Data folders.R
\name{Bundle_RBA_Sheets}
\alias{Bundle_RBA_Sheets}
\title{Reading in RBA Spreadsheets}
\usage{
Bundle_RBA_Sheets(DropboxDataFolder = NULL, ListName = NULL,
  InclVarMeta = FALSE)
}
\arguments{
\item{DropboxDataFolder}{Address for folder containing RBA file locations with file paths saved as the object 'file_path' (and other meta information) created by the Bundle_RBA_Sheets function}

\item{ListName}{Names for the output list}

\item{InclVarMeta}{Logical: include the variable information as columns in working_tbl.}
}
\value{
a list of lists of data frames
}
\description{
Reading in RBA Spreadsheets
}
