% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reading ABS Data folders - Bundle_Sheets.R
\name{Bundle_Sheets}
\alias{Bundle_Sheets}
\title{Reading in ABS Spreadsheets}
\usage{
Bundle_Sheets(DropboxDataFolder = NULL, ListName = NULL,
  InclVarMeta = FALSE, LookForIndustry = FALSE)
}
\arguments{
\item{DropboxDataFolder}{Address for folder containing ABS file locations with file paths saved as the object 'file_path' (and other meta information) created by the Bundle_RBA_Sheets function}

\item{ListName}{Names for the output list}

\item{InclVarMeta}{Logical: include the variable information as columns in working_tbl.}

\item{LookForIndustry}{Logical: enables the process to look for possible columns and row variables which define the industry for the variable from which it creates another column vector on working_tbl with industry abbreviations as how they conform to ANZIC categories.}
}
\value{
a list of lists of data frames
}
\description{
Reading in ABS Spreadsheets
}
