% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSP.R
\name{calc_p}
\alias{calc_p}
\title{Calculate P-value}
\usage{
calc_p(MCsim, test_spd)
}
\arguments{
\item{MCsim}{output from monte_carlo_lsp}

\item{test_spd}{spd to convert to p-value}
}
\description{
Calculates p_value using monte carlo simulation data and a sample spectral power density value.
}

