% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{clean_names_full}
\alias{clean_names_full}
\title{Clean ABS variable names (full)}
\usage{
clean_names_full(names_vec)
}
\arguments{
\item{names_vec}{Character vector (eg. \code{colnames(df)})}
}
\value{
A character vector
}
\description{
Clean ABS variable names (full)
}
