% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download known sheets.R
\name{download_sheets}
\alias{download_sheets}
\title{Collect Multiple ABS Data Sets}
\usage{
download_sheets(url_ls = lsp:::abs_urls)
}
\arguments{
\item{url_ls}{This is a list from within the lsp package of all the ABS URLs that are to be retrieved}
}
\value{
Saves xls files of the spreadsheets to the lsp package folder
}
\description{
Collect Multiple ABS Data Sets
}
