% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Series Graph Format.R
\name{FormArrange}
\alias{FormArrange}
\title{Graph Format templates}
\usage{
FormArrange(GGOutput = NULL, InclColGuide = FALSE, InclFillGuide = FALSE,
  title = NULL, subtitle = NULL, Ytitle = NULL, Ytitle2 = NULL,
  caption = NULL, InclLogo = FALSE)
}
\value{
Image output of arranged components of a graph
}
\description{
Graph Format templates
}
