% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Series Graph Format.R
\name{GplotFormTS}
\alias{GplotFormTS}
\title{Graph Format template 2}
\usage{
GplotFormTS(Facet = FALSE, Yaxis = "", YBreaks = 0, yintercept0 = FALSE,
  XyrBreaks = NA)
}
\arguments{
\item{Facet}{add desc here}

\item{Yaxis}{add desc here}

\item{YBreaks}{add desc here}

\item{yintercept0}{add desc here}

\item{XyrBreaks}{add desc here}
}
\value{
GGplot format
}
\description{
Graph Format template 2
}
