% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphSourcesfn.R
\name{GraphSourcesfn}
\alias{GraphSourcesfn}
\title{Return a table with references for data on each table}
\usage{
GraphSourcesfn(TableName = NULL, Year = NA, Table_No = NA,
  MasterList = TRUE, Data_List = master_list, SubTable = NA,
  Source = NULL, Catalogue = NA, LongTableName = NA)
}
\arguments{
\item{TableName}{add desc here}

\item{Source}{Organisation which is the source of the data set}

\item{Catalogue}{If there is a catelogue identifier write it in here}

\item{LongTableName}{Long name of the data table, with spaces or underscores}

\item{FromMasterList}{Logical: True if Table is in the master list}

\item{MasterTbl}{Name of the master_list table}
}
\value{
Data frame
}
\description{
Return a table with references for data on each table
}
