% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{image_fixer}
\alias{image_fixer}
\title{Image fixture for RMarkdown}
\usage{
image_fixer(path, fig_width_cm = NULL)
}
\arguments{
\item{path}{Character, file path to img as string.}

\item{fig_width_cm}{Numeric, number in cm of width desired.}
}
\description{
Image fixture for RMarkdown
}
