% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_clean.R
\name{meta_clean}
\alias{meta_clean}
\title{Clean ABS Meta Data}
\usage{
meta_clean(meta_data = NULL)
}
\arguments{
\item{meta_data}{This is the ABS excel title page that you read in with the meta data on it.}
}
\value{
A clean dataframe with the meta data.
}
\description{
Clean ABS Meta Data
}
