% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSP.R
\name{monte_carlo_lsp}
\alias{monte_carlo_lsp}
\title{LSP + Monte Carlo}
\usage{
monte_carlo_lsp(t, Nruns, ofac, N_cores = 1)
}
\arguments{
\item{t}{Numeric vector of timepoints}

\item{Nruns}{number of monte carlo iterations}

\item{ofac}{oversampling factor, should be same as in lsp.}

\item{N_cores}{number of cores to use. Default 1.}
}
\value{
Returns raw data of the monte carlo runs. To be passed on to the "MCsim" parameter of function "lsp".
}
\description{
Runs the LSP algorithm a number of iterations using random values on the specified timepoints. Records highest peak at each iteration.
}

