% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reading RBA Data folders.R
\name{process_RBA_sheets}
\alias{process_RBA_sheets}
\title{Processing RBA Spreadsheets into Usable Lists}
\usage{
process_RBA_sheets(input_ls, InclVarMeta = FALSE)
}
\arguments{
\item{input_ls}{List of RBA file locations with file paths saved as the object 'file_path' (and other meta information) created by the Bundle_RBA_Sheets function}

\item{InclVarMeta}{Logical: include the variable information as columns in working_tbl.}
}
\value{
a data frame
}
\description{
Processing RBA Spreadsheets into Usable Lists
}
