% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outmoded Functions.R
\name{process_sheet_old}
\alias{process_sheet_old}
\title{Return data frame with the meta data removed}
\usage{
process_sheet_old(input_ls)
}
\arguments{
\item{input_ls}{ABS data frame}
}
\value{
a data frame
}
\description{
For a typcial sheet 2
}
