% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master process_sheets.R
\name{process_sheets}
\alias{process_sheets}
\title{Return data frame with the meta data removed}
\usage{
process_sheets(input_ls, InclVarMeta = FALSE, LookForIndustry = FALSE)
}
\arguments{
\item{input_ls}{List of ABS file locations with file paths saved as the object 'file_path' (and other meta information) created by the Bundle_Sheets function}

\item{InclVarMeta}{Logical: include the variable information as columns in working_tbl.}

\item{LookForIndustry}{Logical: enables the process to look for possible columns and row variables which define the industry for the variable from which it creates another column vector on working_tbl with industry abbreviations as how they conform to ANZIC categories.}
}
\value{
a data frame
}
\description{
Return data frame with the meta data removed
}
