% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_folder_setup.R
\name{proj_primer}
\alias{proj_primer}
\title{Initiate New Project Folders}
\usage{
proj_primer(Directory = UserDir)
}
\arguments{
\item{Directory}{This is the user directory that has already been generated by the paths package.}
}
\value{
Creates a new set of directories in the new project directory with included YSIcss.css and YSI.png files.
}
\description{
This function creates the standard set of project folders that are needed in each project.
}
