% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphSourcesfn.R
\name{RefCaption}
\alias{RefCaption}
\title{Caption reference}
\usage{
RefCaption(TableNames = NULL)
}
\arguments{
\item{TableNames}{Loaded tables in GraphSources table}

\item{GraphSources}{Reference table made by GraphSourcesfn when loading data}
}
\value{
String
}
\description{
Caption reference
}
