% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outmoded Functions.R
\name{remove_meta}
\alias{remove_meta}
\title{Return data frame with the meta data removed}
\usage{
remove_meta(df)
}
\arguments{
\item{df}{ABS data frame}
}
\value{
a data frame
}
\description{
For a typcial sheet 2
}
