% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminary functions for downloading known
%   sheets.R
\name{tbl_meta}
\alias{tbl_meta}
\title{Forumalte Important Meta Data from an ABS Data Set}
\usage{
tbl_meta(url_ls = lsp:::abs_urls)
}
\arguments{
\item{url_ls}{This is a list from within the lsp package of all the ABS URLs that are to be retrieved and read}
}
\value{
A Data list of dataframes of the meta-data from each ABS data set
}
\description{
Forumalte Important Meta Data from an ABS Data Set
}
