% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outmoded Functions.R
\name{VarNameBreakdownfn}
\alias{VarNameBreakdownfn}
\title{Variable name breakdown}
\usage{
VarNameBreakdownfn(datatable)
}
\arguments{
\item{datatable}{Output from bundle_sheets}
}
\value{
A gathered dataframe with variable names now split by underscore between VariableClass columns
}
\description{
Variable name breakdown
}
