% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Series Graph Format.R
\name{GFormTS}
\alias{GFormTS}
\title{Graph Format templates}
\usage{
GFormTS(Facet = FALSE, Yaxis = "", YBreaks = 0, yintercept0 = TRUE)
}
\value{
GGplot format
}
\description{
Graph Format templates
}
