% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Series Graph Format.R
\name{GplotFormTS}
\alias{GplotFormTS}
\title{Graph Format template 2}
\usage{
GplotFormTS(YBreaks = NA, XyrBreaks = NA, Facet = FALSE, Index = FALSE,
  yintercept0 = FALSE, ExpandToMargins = FALSE, TextSize = 16)
}
\arguments{
\item{YBreaks}{Number of Y-axis breaks (as a guide for pretty_breaks)}

\item{XyrBreaks}{Distance between year breaks}

\item{Facet}{Logical. True if the plot is facetted. Sets the Year breaks to be further apart.}

\item{Index}{Place a horizontal line at height 100 for index measures}

\item{yintercept0}{Place a horizontal line at the zero point}

\item{ExpandToMargins}{Set distance between data points and axes to zero by setting expand = c(0, 0)}

\item{TextSize}{Size of axis labels (uses base_size inside theme_classic)}
}
\value{
GGplot format
}
\description{
Graph Format template 2
}
