% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphSourcesfn.R
\name{GraphSourcesfn}
\alias{GraphSourcesfn}
\title{Return a table with references for data on each table}
\usage{
GraphSourcesfn(TableName = NULL, Data_List = master_list, SubTable = NA,
  Year = NA, Table_No = NA, MasterList = TRUE, Source = NULL,
  Catalogue = NA, LongTableName = NA)
}
\arguments{
\item{TableName}{Working data.frame name for data}

\item{Data_List}{Name of the master_list table}

\item{SubTable}{Table withing Data_List working table is pulled from}

\item{Year}{Year of data release}

\item{Table_No}{Table number}

\item{MasterList}{Logical: True if Table is in the master list}

\item{Source}{Organisation which is the source of the data set}

\item{Catalogue}{If there is a catelogue identifier write it in here}

\item{LongTableName}{Long name of the data table, with spaces or underscores}
}
\value{
Data frame
}
\description{
Return a table with references for data on each table
}
