% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSP.R
\name{lsp}
\alias{lsp}
\title{Lomb Scargle Periodogram}
\usage{
lsp(t, y, ofac = 2, MCsim = NULL, zero_factor = 0, pval_thresh = 1)
}
\arguments{
\item{t}{Numeric vector of timepoints.}

\item{y}{Numeric vector of values corresponding to t.}

\item{ofac}{Oversampling factor. Recommend 2-4. Higher improves granularity, but may cause artefacts.}

\item{MCsim}{MCsim object returned by function "monte_carlo_lsp". If not NULL will return p-values.}

\item{zero_factor}{Integer, zero padding factor. Pads fourier series to increase resolution of approximated function. Number of zeros is equal to zero_factor*length of (unpadded) fourier series. Value of 0 means no padding.}

\item{pval_thresh}{Filter fourier series based on peak strength p value. This will affect the reconstructed series.}
}
\value{
A list with the following elements:\cr
t - original t.\cr
y - original y.\cr
t2 - approximated t.\cr
y2 - reconstructed y. plot against t2.\cr
frequency - frequency range used.\cr
spectral_power_density - Plot against Frequency for the periodogram\cr
peak_info - data.frame with index, frequency, spd, and (optional) pvalue of each detected peak.\cr
peak_spd - highest spd\cr
peak_period - period corresponding to peak_spd\cr
fourier - fourier series\cr
fourier_freq frequencies for fourier series\cr
}
\description{
Calculates the Lomb Scargle Periodogram using the algorithm described by K. Hocke and N. Kämpfer (2009).
}
\references{
Hocke, K., and N. Kämpfer. "Gap filling and noise reduction of unevenly sampled data by means of the Lomb-Scargle periodogram." Atmospheric Chemistry and Physics 9.12 (2009): 4197-4206.
}

