% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_folder_setup.R
\name{proj_primer}
\alias{proj_primer}
\title{Initiate New Project Folders}
\usage{
proj_primer(UserDir = NULL, proj_path = NULL, proj_name = NULL)
}
\arguments{
\item{UserDir}{Character vector of the user's primary drive dile pather, e.g C:/User/Users}

\item{proj_path}{Character vector of the directory where the new project will be housed, e.g Documents/R/Projects}

\item{proj_name}{Character vector of the name of the new project you are creating}
}
\value{
Creates a new set of directories in the desired location with included YSIcss.css and YSI.png files
}
\description{
Initiate New Project Folders
}
